DB リネーム

+++ client/src/app/home/home.component.ts
@@ -5,6 +5,7 @@ import { ActivatedRoute, ParamMap } from '@angular/router';
 import { MatDialog } from '@angular/material/dialog';
 
 import { environment } from '../../environments/environment';
+import { DbRenameDialogComponent } from './db-rename-dialog.component';
 import { DbDeleteConfirmDialogComponent } from './db-delete-confirm-dialog.component';
 
 /** Home : DB 一覧画面 */
@@ -41,7 +42,22 @@ export class HomeComponent implements OnInit {
   }
   
   /**
-   * DB を削除するか確認のダイアログをを表示する
+   * DB をリネームするダイアログを表示する
+   * 
+   * @param dbName リネーム対象となる DB 物理名
+   */
+  public openRenameDialog(dbName: string): void {
+    const dialogRef = this.matDialog.open(DbRenameDialogComponent, {
+      data: { dbName }
+    });
+    dialogRef.afterClosed().subscribe(async (dialogData) => {
+      if(!dialogData || !dialogData.isRenameConfirmed) return console.log('DB Rename Dialog Was Closed. Nothing To Do');
+      await this.renameDb(dbName, dialogData.newDbName);
+    });
+  }
+  
+  /**
+   * DB を削除するか確認のダイアログを表示する
    * 
    * @param dbName 削除対象となる DB 物理名
    */
@@ -49,8 +65,8 @@ export class HomeComponent implements OnInit {
     const dialogRef = this.matDialog.open(DbDeleteConfirmDialogComponent, {
       data: { dbName }
     });
-    dialogRef.afterClosed().subscribe(async (isDeleteConfirmed) => {
-      if(!isDeleteConfirmed) return console.log('DB Delete Confirm Dialog Was Closed. Nothing To Do');
+    dialogRef.afterClosed().subscribe(async (dialogData) => {
+      if(!dialogData || !dialogData.isDeleteConfirmed) return console.log('DB Delete Confirm Dialog Was Closed. Nothing To Do');
       await this.deleteDb(dbName);  // DB を削除し一覧を再描画する
     });
   }
@@ -76,7 +92,31 @@ export class HomeComponent implements OnInit {
   }
   
   /**
-   * DB を削除し一覧を再描画する
+   * DB をリネームし一覧を再取得する
+   * 
+   * @param dbName リネーム対象の DB 物理名
+   * @param newDbName リネーム後の DB 物理名
+   */
+  private async renameDb(dbName: string, newDbName: string): Promise<void> {
+    this.errorMessage = '';
+    this.isLoading = true;  // `isSubmitting` は省略する…
+    try {
+      const result: any = await this.httpClient.put(`${environment.apiRootPath}/db/${dbName}/name`, { dbName, newDbName }).toPromise();
+      console.log('Rename DB : Success', result);
+      this.successMessage = 'DB Renamed.';
+      this.isHideSuccessMessage = false;
+      setTimeout(() => { this.isHideSuccessMessage = true; }, 4000);
+      await this.getDbList();  // DB リストを再取得する
+    }
+    catch(error) {
+      console.error('Rename DB : Failed', error);
+      this.errorMessage = error.error?.error || error.error || error.toString();
+      this.isLoading = false;
+    }
+  }
+  
+  /**
+   * DB を削除し一覧を再取得する
    * 
    * @param dbName 削除対象の DB 物理名
    */
